Seperate files for server side paginated tables
