+++ src/app/admin/usersTable/constants.ts
@@ -1,0 +1,2 @@
+export const defaultNumberOfUsersPerPage = 10;
+export const numberOfUsersPerPageOptions = [10, 25, 50, 100];
