+++ src/app/clients/clientsTable/constants.ts
@@ -1,0 +1,1 @@
+export const clientIdParam = "clientId";
