+++ src/app/parcels/ActionBar/StatusesModal.tsx
@@ -4,9 +4,9 @@ import Button from "@mui/material/Button/Button";
 import Modal from "@/components/Modal/Modal";
 import { DatePicker, TimePicker } from "@mui/x-date-pickers";
 import dayjs, { Dayjs } from "dayjs";
-import { ParcelsTableRow } from "@/app/parcels/getParcelsTableData";
 import SelectedParcelsOverview from "./SelectedParcelsOverview";
 import { ErrorSecondaryText } from "@/app/errorStylingandMessages";
+import { ParcelsTableRow } from "../parcelsTable/types";
 
 interface StatusesBarModalProps extends React.ComponentProps<typeof Modal> {
     selectedParcels: ParcelsTableRow[];
