+++ src/app/parcels/parcelsTable/constants.ts
@@ -1,0 +1,5 @@
+export const parcelIdParam = "parcelId";
+
+export const defaultNumberOfParcelsPerPage = 100;
+
+export const numberOfParcelsPerPageOptions = [10, 25, 50, 100];
