+++ src/pdf/DayOverview/DayOverviewPdfButton.tsx
@@ -7,7 +7,7 @@ import PdfButton from "@/components/PdfButton/PdfButton";
 import DayOverviewPdf from "./DayOverviewPdf";
 import { logErrorReturnLogId } from "@/logger/logger";
 import { PdfDataFetchResponse } from "../common";
-import { ParcelsTableRow } from "@/app/parcels/getParcelsTableData";
+import { ParcelsTableRow } from "@/app/parcels/parcelsTable/types";
 
 interface Props {
     onPdfCreationCompleted: () => void;
